requirements.txt
python-dotenv
httpx[http2]==0.27.0
pydantic==2.7.0
python-ntfy
